commit latest java code

+++ src/main/java/com/example/flash2/entity/CourseFeeEntity.java
@@ -29,6 +29,6 @@ public class CourseFeeEntity {
   private Integer CourseTypeId;
 
   @Column(name = "FEE_PER_STUDENT")
-  private float FeePerStudent;
+  private double FeePerStudent;
 }
